moved hangman.py to server.py
